+++ src/pages/Home.jsx
@@ -2,8 +2,9 @@ import React, { useEffect } from 'react';
 import Navbar from '../components/Navbar/Navbar';
 import Hero from '../components/Hero/Hero';
 // import About from '../components/About/About';
-import Services from '../components/Services/Services';
-import MyWork from '../components/MyWork/MyWork';
+import Skillset from '../components/Skillset/Skillset';
+import Socials from '../components/Socials/Socials';
+import Projects from '../components/Projects/Projects';
 import Contact from '../components/Contact/Contact';
 import Footer from '../components/Footer/Footer';
 
@@ -21,8 +22,9 @@ const Home = () => {
       </header>
       <main>
         <Hero />
-        <Services />
-        <MyWork />
+        <Skillset />
+        <Projects />
+        <Socials />
         <Contact />
       </main>
       <Footer />