+++ src/components/Navbar/Navbar.jsx
@@ -5,11 +5,10 @@ import './Navbar.scss';
 import logo from '../../assets/global/header-logo-default.png';
 import menu_open from '../../assets/menu_open.svg';
 import menu_close from '../../assets/menu_close.svg';
-import underline from '../../assets/nav_underline.svg';
 
-const Navbar = () => {
+const Navbar = ({ isHome = false }) => {
   
-  const [menu,setMenu] = useState("home");
+  //const [menu,setMenu] = useState("home");
   const menuRef = useRef();
 
   const openMenu = () => {
@@ -31,6 +30,11 @@ const Navbar = () => {
                 {/* Keep old link just in case */}
                 {/* <li><AnchorLink className='anchor-link' href='#home'><p onClick={()=>setMenu("home")}>Home</p></AnchorLink>{menu==="home" ? <img src={underline} alt='' />:<></>}</li> */ }
                 <li><NavLink className={({ isActive }) => isActive ? "anchor-link current" : "anchor-link" } to="/">Home</NavLink></li>
+                { isHome ? (
+                  <li><AnchorLink className='anchor-link' offset={0} href='#projects'>Projects</AnchorLink></li>
+                ) : (
+                  <li><NavLink className='anchor-link' to="/#projects">Projects</NavLink></li>
+                ) }
                 <li><NavLink className={({ isActive }) => isActive ? "anchor-link current" : "anchor-link" } to="/about">About</NavLink></li>
                 <li><NavLink className={({ isActive }) => isActive ? "anchor-link current" : "anchor-link" } to="/twitch">Twitch</NavLink></li>
                 {/* <li><AnchorLink className='anchor-link' offset={50} href='#about'><p onClick={()=>setMenu("about")}>About Me</p></AnchorLink>{menu==="about" ? <img src={underline} alt='' />:<></>}</li>
@@ -38,7 +42,11 @@ const Navbar = () => {
                 <li><AnchorLink className='anchor-link' offset={50} href='#mywork'><p onClick={()=>setMenu("mywork")}>Portfolio</p></AnchorLink>{menu==="mywork" ? <img src={underline} alt='' />:<></>}</li>
                 <li><AnchorLink className='anchor-link' offset={50} href='#contact'><p onClick={()=>setMenu("contact")}>Contact</p></AnchorLink>{menu==="contact" ? <img src={underline} alt='' />:<></>}</li> */}
             </ul>
-            <div className="nav-connect"><AnchorLink className='anchor-link' offset={50} href='#contact'>Connect</AnchorLink></div>
+            { isHome ? (
+              <div className="nav-connect"><AnchorLink className='anchor-link' offset={0} href='#contact'>Connect</AnchorLink></div>
+            ) : (
+              <div className="nav-connect"><NavLink className='anchor-link' to='/#contact'>Connect</NavLink></div>
+            )}
         </nav>
       </div>
     </header>