+++ CHANGELOG.md
@@ -16,10 +16,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [Twitch] Embedded player
 - [Twitch] History
 - [Twitch] Now Streaming + Community
+- [Projects] added
+- [Socials] added
 ### Changed
 - [Global] Cleaned up global styles
 - [Global] Fixed pages to top scroll for each load
 - [About] Updated/cleaned layout
+- [Home] Updated layout
+### Deprecated
+- [Services] removed
 
 ## [0.2.4] - 2024-08-28
 ### Added